delete authenticator service
